+++ src/test/java/subway/station/StationAcceptanceTest.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.station;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
@@ -10,6 +10,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
+import subway.station.domain.Station;
+import subway.station.repository.StationRepository;
 
 import java.util.HashMap;
 import java.util.List;
